Added support for additional Aviary config options
Added the ability to supply extra Aviary configuration options in an
aviary_options array and moved configuration of existing Aviary options
to the new format (original format will still work for old config
files).

Also removed the lower casing of file names in fix_filename when
convert spaces is enabled- this was breaking saving of files with
uppercase letters and it’s unclear why it’s even part of the convert
spaces functionality.